python/requirements.txt
py3exiv2
arrow
python-magic
